test: Lengthen the delay in USDT defer trigger.
The trigger usdt-tst-defer.c expects that, as soon as probe foo is
enabled, so too are the foo and bar is-enabled probes as well as
the bar probe.  Since they are not turned on absolutely concurrently,
the trigger actually has a microsleep after foo is enabled and
before further activity is resumed.  Nonetheless, there is evidence
in occasional test runs that some events are missed.

Since there are no guarantees how concurrently these probes are
turned on, lengthen the delay.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>