test: Improve sync mechanism in tst.multiple.sh
In this test, a USDT process is launched.  Subsequently, dtrace is
started.  To give dtrace a chance to enable the USDT probes, the
USDT process waits a few seconds.  On some systems and when the
amount of work dtrace does upon startup is high, "a few seconds"
may not be enough, causing the test to miss expected probes.

Have the USDT process loop on an is-enabled probe;  this is a
more resilient sync mechanism between it and dtrace.

That said, dtrace does not enable all probes in one instant, and
it makes no promises in what order probes will be started.  So,
we still leave a (shorter) sleep() in the target process.

Arguably, the test could be simplified by launching the USDT
process with "dtrace -c".  In this patch, we choose instead to
retain the overall character of the existing test.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>